+++ client/fingerprint/network_default.go
@@ -1,0 +1,8 @@
+// +build !linux
+
+package fingerprint
+
+// linkSpeed returns the default link speed
+func (f *NetworkFingerprint) linkSpeed(device string) int {
+	return 0
+}
